Temporal edited version of SV-OR

+++ src/server.c
@@ -8,7 +8,7 @@
 #include "project.h"
 
 extern int shmid;
-extern int msgid[4];
+extern int msgid[NODENUM];
 extern int id;//in common.c, id for server
 extern int mode;//from project.c, indicate client/server-oriented mode
 
@@ -20,9 +20,9 @@ static const char *dat[4] = {
 };
 static int fd = -1;
 static int* shm_addr = NULL;
-static int ser_buf[8] = {0, };
+static int ser_buf[CHKSIZE] = {0, };
 static pid_t parent;
-
+static int sv_shut;
 
 static void
 read_chunk_shm(int sig)
@@ -48,12 +48,7 @@ shutdown(int sig)
 
 		else if(mode == MODE_SVOR)
 		{
-				close(fd);
-				if (msgctl(msgid[id], IPC_RMID, NULL) == -1)
-				{
-       					perror("msgctl");
-        				exit(-1);
-				}
+				sv_shut = 1;
 		}
 
 		exit(0);
@@ -96,24 +91,29 @@ do_server_task(int mode)
 		else if(mode == MODE_SVOR)
 		{
 		    static msgbuf msg;
-			static sigset_t mask;
-			sigemptyset(&mask);
-			sigaddset(&mask, SIGUSR1);
 			while(1)
 			{
-				sigsuspend(&mask);  // wait until msg queue is full, signal by parent
 				for (int i = 0; i < 8; i++)
-				{	
-					nbyte = msgrcv(msgid[id], &msg, sizeof(int), -7, 0);
+				{
+					nbyte = msgrcv(msgid[id], &msg, sizeof(int), -8, 0);
 					if (nbyte == -1)
 					{
 						perror("msgrcv");
 						exit(-1);
 					}
-					ser_buf[msg.mtype] = msg.mtext[0];
+					ser_buf[msg.mtype - 1] = msg.mtext[0];
 				}
 				write(fd, ser_buf, CHKSIZE);  // write data to file
-				kill(parent, SIGUSR2);
+				if (sv_shut)
+				{
+					close(fd);
+					msgctl(msgid[id], IPC_RMID, NULL);
+					break;
+				}
+				else
+				{
+					kill(parent, SIGUSR2);
+				}
 			}
 		}
 		exit(0);//no return!, do_client_task is called inside of for-loop with fork() common.c:24
